requirements.txt
asgiref==3.2.10
backcall==0.2.0
beautifulsoup4==4.9.1
certifi==2020.6.20
chardet==3.0.4
decorator==4.4.2
Django==3.1
django-filter==2.3.0
djangorestframework==3.11.1
idna==2.10
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.17.2
lxml==4.5.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.6
ptyprocess==0.6.0
Pygments==2.6.1
pytz==2020.1
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
traitlets==4.3.3
urllib3==1.25.10
wcwidth==0.2.5
